+++ src/js/constants.js
@@ -68,6 +68,7 @@ export const EVENT_SCALED = 'scaled';
 export const EVENT_ZOOM = 'zoom';
 export const EVENT_ZOOMED = 'zoomed';
 export const EVENT_PLAY = 'play';
+export const EVENT_PLAYING = 'playing';
 export const EVENT_STOP = 'stop';
 
 // Data keys